# docs/swagger.yaml
basePath: /
definitions:
...
  databaseTypes.LostAndFound:
    properties:
      date_found:
        example: "2022-01-01T12:00:00Z"
        type: string
      description:
        type: string
      id:
        example: 1
        type: integer
      image_file:
        items:
          type: integer
        type: array
      image_url:
        type: string
      item_name:
        example: Backpack
        type: string
      location_found:
        example: Library
        type: string
      status:
        example: 1
        type: integer
      submitter_id:
        example: 2
        type: integer
    type: object
...
  lostAndFound.deleteResponse:
    properties:
      status:
        type: string
    type: object
...
  restTypes.LostAndFoundErrorResponse:
    properties:
      error:
        type: string
    type: object
  restTypes.LostAndFoundPostResponse:
    properties:
      id:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
...
info:
  contact:
    name: Senya
  description: Simple swagger implementation in Go HTTP
  title: Go Rest API with Swagger for school system
  version: "1.0"
paths:
...
  /data/lost-and-found/:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all lost and found items from the database
        and returns them as a JSON array in the response body.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/databaseTypes.LostAndFound'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all lost and found items
      tags:
      - LostAndFound
    post:
      consumes:
      - multipart/form-data
      description: Adds a lost and found item to the database and uploads an image
        file
      parameters:
      - description: Name of the lost/found item
        in: formData
        name: item_name
        required: true
        type: string
      - description: Description of the lost/found item
        in: formData
        name: description
        type: string
      - description: Date the item was found
        in: formData
        name: date_found
        required: true
        type: string
      - description: Location where the item was found
        in: formData
        name: location_found
        required: true
        type: string
      - description: Status of the item (lost or found)
        in: formData
        name: status
        required: true
        type: string
      - description: Image of the lost/found item
        in: formData
        name: image_file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/restTypes.LostAndFoundPostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/restTypes.LostAndFoundErrorResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/restTypes.LostAndFoundErrorResponse'
      security:
      - Bearer: []
      summary: Add a lost and found item
      tags:
      - LostAndFound
  /data/lost-and-found/image/{id}:
    put:
      consumes:
      - application/json
      description: Update an existing lost and found item in the database with the
        specified ID
      operationId: update-lost-and-found-item
      parameters:
      - description: Lost and found item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Item name
        in: formData
        name: item_name
        type: string
      - description: Item description
        in: formData
        name: description
        type: string
      - description: Date the item was found
        in: formData
        name: date_found
        type: string
      - description: Location where the item was found
        in: formData
        name: location_found
        type: string
      - description: Status of the lost and found item
        in: formData
        name: status
        type: string
      - description: Image file of the lost and found item
        in: formData
        name: image_file
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/restTypes.LostAndFoundErrorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/restTypes.LostAndFoundErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/restTypes.LostAndFoundErrorResponse'
      security:
      - Bearer: []
      summary: Update a lost and found item
      tags:
      - LostAndFound
  /data/lost-and-found/image/{imageID}:
    get:
      consumes:
      - application/json
      description: Fetches the image file for a lost and found item with the specified
        ID from the database and returns it as a JPEG image in the response body.
      parameters:
      - description: The ID of the lost and found item to retrieve the image file
          for.
        in: path
        name: imageID
        required: true
        type: string
      produces:
      - image/jpeg
      responses:
        "200":
          description: The image file for the specified lost and found item.
          schema:
            type: string
        "404":
          description: The specified lost and found item ID was not found in the database.
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get the image file for a lost and found item by ID.
      tags:
      - LostAndFound
...
  /lost-and-found/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a lost and found item from the database
      operationId: delete-lost-and-found-item
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lostAndFound.deleteResponse'
        "400":
          description: Invalid item ID
          schema:
            type: string
        "404":
          description: Item not found
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete a lost and found item
      tags:
      - LostAndFound
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"